Error handling for contexts where path_and_mimetype() will fail
