+++ docassemble/ALWeaver/custom_values.py
@@ -24,7 +24,8 @@ Container for widely used data that may be altered by user inputs.
 """
 
 __all__ = ['get_possible_deps_as_choices', 'get_pypi_deps_from_choices', 
-           'get_yml_deps_from_choices', 'SettingsList', 'load_capabilities']
+           'get_yml_deps_from_choices', 'SettingsList', 'load_capabilities',
+          'advertise_capabilities']
 
 class SettingsList(DAList):
   """
@@ -49,13 +50,15 @@ def load_capabilities(base:str="docassemble.ALWeaver", minimum_version="1.5", in
   this_yaml = path_and_mimetype("data/sources/configuration_capabilities.yml")[0]
   weaverdata = DAStore(base=base)
   published_configuration_capabilities = weaverdata.get("published_configuration_capabilities") or {}
-  
-  with open(this_yaml) as f:
-    this_yaml_contents = f.read()
+  try:
+    with open(this_yaml) as f:
+      this_yaml_contents = f.read()
 
-  first_file = list(yaml.safe_load_all(this_yaml_contents))[0]
-  
-  capabilities = {"Default configuration": first_file}
+    first_file = list(yaml.safe_load_all(this_yaml_contents))[0]
+
+    capabilities = {"Default configuration": first_file}    
+  except:
+    capabilities = {}
   
   for key in list(published_configuration_capabilities.keys()):
     # Filter configurations based on minimum published version
@@ -126,3 +129,25 @@ def get_yml_deps_from_choices(choices:Union[List[str], DADict]):
   else: # List
     return choices
     
+def _package_name():
+  """Get package name without the name of the current module, like: docassemble.ALWeaver instead of
+  docassemble.ALWeaver.advertise_capabilities"""
+  try:
+    return ".".join(__name__.split(".")[:-1])
+  except:
+    return __name__
+
+def advertise_capabilities(package_name:str=None, yaml_name:str="configuration_capabilities.yml", base:str="docassemble.ALWeaver", minimum_version="1.5"):
+  """
+  Tell the server that the current Docassemble package contains a configuration_capabilities.yml file with settings that
+  ALWeaver can use, by adding an entry to the global DAStore. This function should be set to run with a 
+  # pre-load hook so it advertises itself on each server uwsgi reset.
+  """
+  weaverdata = DAStore(base=base)
+  if not package_name:
+    package_name = _package_name()    
+  published_configuration_capabilities = weaverdata.get("published_configuration_capabilities") or {}
+  if not isinstance(published_configuration_capabilities, dict):
+    published_configuration_capabilities = {}
+  published_configuration_capabilities[package_name] = [yaml_name, minimum_version]
+  weaverdata.set('published_configuration_capabilities', published_configuration_capabilities)
